+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Recent changes
 - NEW: Cmdlets now return true or false, depending on the results.  Thanks to @o-l-a-v for suggesting this in issue #1.
 
+## Version 1.2.0 (January 18, 2024)
+- NEW: Add support for Microsoft's [new MX records that support DANE and DNSSEC](https://techcommunity.microsoft.com/t5/exchange-team-blog/implementing-inbound-smtp-dane-with-dnssec-for-exchange-online/ba-p/3939694).  Note that this does not go live until March 2024 (as a preview), so there may be bugs that we don't yet know about.
+- NEW: `Test-ExchangeOnlineMxRecords`, `Test-ExchangeOnlineRecords`, and `Test-Office365DnsRecords` now support a new `-DANERequired` parameter that prints a warning if the DANE-enabled MX endpoint is *not* in use.
+
 ## Version 1.1.0 (June 1, 2023)
 - NEW: Add support for checking DNS records for Office 365 hosted by 21Vianet.
 - ENHANCEMENT: The Azure AD client configuration record check now passes if the `msoid` record is missing.  As of <time datetime="2023-02-16">early 2023</time>, this DNS record is only required to be set when Office 365 is managed by 21Vianet.